+++ app/components/ors-header/ors-header.js
@@ -44,8 +44,8 @@ angular.module("orsApp.ors-header", []).component("orsHeader", {
         ctrl.lists_extra_info = lists.extra_info;
         ctrl.getActiveProfile = orsSettingsFactory.getActiveProfile;
         ctrl.getActiveSubGroup = () => {
-          return lists.profiles[ctrl.getActiveProfile().type].subgroup
-        }
+          return lists.profiles[ctrl.getActiveProfile().type].subgroup;
+        };
         ctrl.optionList = lists.userOptions;
         ctrl.changeExtras();
       };