+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - deviation tolerance during route import ([#325](https://github.com/GIScience/openrouteservice-app/issues/325))
     - set lower value for more waypoints/accuracy
 
+### Changed
+- maxZoom of all basemaps to 18 ([#311](https://github.com/GIScience/openrouteservice-app/issues/311))
+
 ## [v0.4.2] - 2020-01-30
 
 ### Added