+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - activate in settings, click route or sidebar entry to switch route
     - only for routes with 2 points (start & end) below 100km
     - default values for `weight-factor` and `share-factor` used
+- possibility to mock the ors response for development purpose
+    - uncomment all instances of "mock" to enable
 
 ### Changed
 - maxZoom of all basemaps to 18 ([#311](https://github.com/GIScience/openrouteservice-app/issues/311))