+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Deprecated
 ### Removed-->
 
-## [Unreleased]
+## [v0.4.3] - 2020-03-26
 
 ### Added
 - deviation tolerance during route import ([#325](https://github.com/GIScience/openrouteservice-app/issues/325))
@@ -24,8 +24,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - possibility to mock the ors response for development purpose
     - uncomment all instances of "mock" to enable
 
+### Fixed
+- no extra info settings for subgroups ([#328](https://github.com/GIScience/openrouteservice-app/issues/328))
+
 ### Changed
-- maxZoom of all basemaps to 18 ([#311](https://github.com/GIScience/openrouteservice-app/issues/311))
+- maxZoom of all base maps to 18 ([#311](https://github.com/GIScience/openrouteservice-app/issues/311))
 - route processing logic to increase performance ([#320](https://github.com/GIScience/openrouteservice-app/issues/320))
 
 ## [v0.4.2] - 2020-01-30